Move files for a better structure Directly read genesis.json and set 'abci_genesis'
